Commit 10/27
I thought I had commited AND pushed to master, but I didn't check and it looks like it was not actually commited. On 10/27, I worked on finishing up the I2C transmission from the gyro/accel to the arduino. In the new code, I set the sensitivities of the gyro to 131 Least Significant Bits per degree per second, and the accelerometer to 16384 LSB/g. These should be the most sensitive settings, if my understanding of LSBs are remotely correct. Afterwards, I request values from the gyro/accel in the main update loop and attempt to print to the serial(console), but for some reason, nothing is printing to the serial.