Commit 10/30
Not sure what was happening with the serial not printing any values from the MPU-6050. I did a bad soldering job on the MPU-6050 so the connection for the power was poor, causing it to turn on and off regularly. Therefore, I focused on getting started with the radio receiver code, initializing the nRF24L01 receiver that is connected to the arduino. The code is set up so that I just need to add additional processing in the if statement. Other than that, I imported the PID library which was published on arduino's official website, which made things easy. I will do more experimentation on the PID later. There is also a PID auto-tuning library that may come handy, since tuning the PID Controller is the hardest/most annoying part of using PID.