Commit 10/31
I started working on the coordinate system where I keep track of the drone heading (yaw, pitch, roll), and the position. Andrew M., Nathan, and I went outside to talk about a problem with tracking position using the accel/gyro. Let's say the drone is moving sideways. This means that the drone won't be parallel to the ground, so even though there is only horizontal movement, the accelerometer will say that there was partial vertical motion, and partial horizontal motion. We worked on the math to compensate for this, knowing the angle of the drone with respect to the ground. We talked until nearly the end of class, so next period, I will work on a code implementation of this.