Commit 11/1
Today, we mainly worked on the math for correcting positional values. We tried using various coordinate systems (spherical, etc) since the gyro returns values for X Y and Z Angular velocities, but since one of them isn't a problem, we tried spherical and the correction "algorithm" that we came up with might work. But we ran into an issue because the gyro always uses the Z axis as the orientation, so we'll see if the code works.