Commit 11/3
I was going to work on putting into code our math that we did last time, but I ran into an interesting resource where the poster was talking about timing and space efficiency. Because arduino is made to be friendly to noobies, some methods aren't as efficient (storage and speed wise) as compared to direct bit I2C communication. So I will look further into that over the weekend.