Commit 11/6
While I was implementing PID, I needed to know the specificity of the time provided by the controller, and how reliable it is. After all, the goal is to have the drone fly as smoothly and stably as possible. Before, I was setting a delay in the main loop, but I found that such solution has A LOT of variance, and also that my previous selection of 100 Hz is not ideal, so now I am tracking time in microseconds, using the micros() function, which has the least delay and jitter.