Commit 11/13
Wrote code for calibrating the ESCs. Normally, expensive wireless transmitters have a mode for calibrating the ESCs, but since we aren't using one, I just made basic code to feed PWM pulses to the ESCs. Afterwards, I just need to connect to the ESC and follow the instructions on the paper manual that came with the ESC.