Commit 11/15
I made a new arduino sketch for integrating gyro values and trying out different senstivities and calibrating the gyro on startup, but for some reason, it won't save to the GitHub folder, so I changed 1 line of code on the quadcopter sketch to make the commit. I will work on getting the sketch to save (may require system restart) and put it in a commit later today. Basically, it was what I implemented in the quadcopter sketch, but I couldn't test it on the quadcopter sketch because there was other code and it would become too difficult if I tested it directly on there. I found that the initial sensitivity for the gyro that I used (250 degrees per second) was too slow and was prone to accumulating error even with reasonable turning speed, so I decided to change it to 500 degrees per second. I am losing a bit of accuracy, but it does mean that the gyro is capable of reading up to 500deg/s which is a reasonable speed when there is sudden motion. I will do the same with the accelerometer and put it back into the quadcopter sketch.