Commit 11/17
I was trying to import a library for calculus so doing functions for gyro and accel would be easier. That didn't work, so I tried finding a library for the MPU-6050, and tried importing it, but that didn't work either. For some reason, I am unable to import any new libraries. Might need to save my work, reinstall the arduino IDE, and re-import all my work.