Commit 11/28
Worked the entire period working with Nathan on figuring out how to make the radio work. For some reason, there is a huge delay whenever we try to send a string from the serial through the radio to the receiver. We tried various combinations of using different string methods in C, but still no luck. Might have to use bytes instead of STrings.