Commit 11/29
Removed some extraneous stuff from my code that I'm going to re-implement more accurately and efficiently later once the basic code is done. For example, I was trying to start implementing the coordinate system for the drone, but without the gyro and pid stuff all done, it was just pointless having it. Today, I worked on some math for PID and gyro calibration. Apparently gyro readings can be affected by the temperature of the sensor? So I looked into that, but I'm not sure how significant that will be. So first, I will create a "vanilla" implementation of everything, with just fundamentals. Depending on the performance, I will add small changes like temperature.