Commit 12/5
I was outside the room for most of the period and forgot to commit as a result. I helped test Andrew's drone, from radio range, and we did a re-calibration of the ESCs because for some reason one of the motors is clearly starting earlier than the others. Even after re-calibration, that was still the case.

We also "tested" the drone's hovering. Don't worry, we didn't actually fly the drone. What we did was, went outside the building, and tried to see if a person holding the drone could feel a force on a drone that would cause it to topple. We found that the drone had a tendency to rotate clockwise, but we do not know why. I suspect that it was because the motors are spinning at a different speed (the 2 clockwise motors might have been running more quickly than the 2 counter-clockwise motors, but we did do the calibration). Perhaps, there is something wrong with the calibration.