Commit 12/6
I might have said before, but there was significant drift in the calculated Z angle, most likely due to gyroscope's inherent drift. But that inherent drift should be around 0.25deg/min, so something else was happening. For some reason, there is almost no drift now. What I noticed was that, when I physically touch the sensor with my hands, the error increases a lot. The issue might be that the temperature and the humidity from my hands are affecting the sensor's readings. Anyway, if it works, then cool.