Commit 12/11
Consolidated gyro code into main quadcopter code. I tried to test the PID using prints, but for some reason the Yaw adjustment is always returning max. Judging from the prints, there is a problem with the sign. I do not think there is a problem with the PID constants. Will think and debug...