Commit 12/12
Forgot to commit again. I am always working to the end of the period, and keep forgetting to commit because I don't want to be late to the next class. Starting today, I will just end work 5 minutes before class. So on this day, I realized that the PID output for Yaw was always stuck to -410, for some reason (also, I just realized that the PID outputs were not being capped to the pid_max value that I set at the top of the file). I'm going to get rid of the yaw output, and see if the ESC output values look a bit better.