Commit 1/8
Re-organized code because it was getting too confused to work with, since everything is on one file. In terms of the actual code, the main focus today was ensuring that the main loop operates at a constant 250Hz or every 4000us. That way, the calculations for angle and PID are more consistent.