Commit 1/9
I mainly helped Andrew outside today. The problem he was having today was similar to mine: the radio is disconnecting occasionally, but it is completely random. I will have to check the wiring for stability. Another problem was that the accelerometer calibration is off. Actually, the accelerometer has natural offset, but if we calibrate it using the starting position, I neeed to make sure the ground is completely flat, otherwise the calculated offset itself will have error.