Commit 1/12
Today, I spent the beginning half of the class helping Nathan and Jeffrey debug their drone code to test propellers on setup and loop to make sure the drone wasn't starting with the propellers getting power. For the second half, I was testing my PID controller again because for some reason the throttle would go down to 0 when it should stay at a constant 1000.