Commit 1/16
Forgot to commit yesterday again. Mainly worked on re-assembling the drone after having re-soldered the connections. Andrew (finally) brought me the mounting tape that I needed to secure the electronics onto the frame. Everything should be rewired properly. Should be. I will be testing the wiring and then do a final manual testing of the PID with gyro/accel and then trying it on the actual drone.