Added Space Invaders Pre-Project
This Space Invaders project is the project that I worked on during the "learning" period of time before the actual project for 2nd semester. Included is a Mac executable .app build of the game. A Windows version is available, but I need to build it separately on Windows, so if you would like that instead, please let me know.