Commit 4/17
Today, I worked on linking input management to the actually tetrimino itself. For some reason, though, when I try rotating the tetriminos, the center of the rotation is not defined, so the tetrimino disappears rather than rotating. It may be because the tetriminos are constructed as a group of 4 cube gameobjects under a singular empty gameobject parent. The parent gameobject, which has the script applied to it, has no physical bounds so it's rotating relative to the center of the world rather than the center of the shape. Other than that, I worked on making the tetrimino moving down at a constant speed.