Commit 4/20
I was further refining the user controls, and one really important key is the spacebar. Pressing the spacebar would make the floating piece drop to where it would fall all the way. The only problem is that I don't know where the "floor" is, since the floor isn't necessary the bottom of the game box, but where the highest point in the path of the current piece is. You also have to account for the shape of the current shape and the shape of the "floor" and figure out which side touches first, and it's a hassle to do it without a grid. Technically a workaround is by making the velocity super fast when the spacebar is pressed, but because of the way collisions are handled (every frame), if the velocity is too fast, the piece might just move through the floor. So I need to make a grid, and that's what I started working on today.