Commit 4/24
Today was a short day. I spent around 40% of the time helping Nathan because his player movement was slow and delayed, whereas the goal was to be as swift and responsive. I helped him fix it. Other than that, I managed to get the continuous motion working. It is quite good, and I will be trying out raycasting for the "drop" feature that I mentioned in the retrospective.