Commit 5/2
Another productive day--I corrected the infinite loop problem (when the game is supposed to end, it used to infinitely generate tetrominos). I also "fixed" pieces moving outside the gameplay box, but it's very buggy because the center of rotation of the tetriminos could mean that even though the center of rotation (its position) may be inside the box, if you rotate it, it may go out of the box. I also implemented the piece preview thing where you can see in advance where the current tetrimino is going to land if you were to drop it.