Commit 5/4
I had to deal with a barrage of questions from my classmates about how collisions work in unity and some other questions about sprites and rigidbodies. Other than that, I spent time working on making sure pieces can't go through eachother when moving sideways. I used the same raycasting approach to see if the minimum distance to the left or the right was at least 1. If not, the movement is not actually done.