Commit 5/8
Today I focused on fixing the bug where if you rotate a tetromino one of its squares may go out of bounds. I tried to deal with this by moving the tetromino horizontally back into the play box, but I ran into a problem where Unity's local position and global position is different for a given transform depending on whether it is part of a parent gameobject or not. I will address this issue next time. I will be absent on Wednesday because I will be taking an AP test.