Commit 5/16
I forgot to press commit and push. I know I won't get credit, but I'll push anyway so you know I wasn't just goofing around on Wednesday. I tried to come up with ways to prevent the pieces rotating out of the gameplay box. The main problem I'm encountering is this: the solution I'm thinking of involves shifting the pieces away from the wall 1 piece if it happens to rotate out of the box. I would first test if it rotates out of bounds and then if it does, then shift the piece. The problem with that is if the piece I'm rotating is already next to another tetromino that's been placed. Then, if I shift the tetromino then it may go into eachother.