Commit 5/22
I was trying to work on the manual collision intersection/detection function. Apparently unity has this feature, but my unity is outdated (version 5.5). Now it's like Unity 2018 or something, which is 2 versions after 5.6, which is when the Physics2D.ComputePenetration method that I need to use came out. So I spent time updating Unity. It still hasn't finished. You can't see that I updated Unity yet because 1, I haven't finished updating, and 2, the gitignore ignores all the files other than files strictly related to the working set of the project.