Commit 5/23
I arrived late in class, and despite it being a Wednesday, today we had a weird schedule due to state testing so class period was short. I finished the update shortly after coming to class, but noticed that Unity had not updated to the latest version. It turns out, the "check for updates" feature on Unity (which I used) only checks for newer updates on the same major version of Unity. So I have to completely install a new version from scratch... I'll do it at home tonight.