Commit 5/29
Made a lot of progress restructuring the game mechanic. Finally decided on using grid implementation as that helps reduce my dependence on raycasting and also creates the foundation to avoid all of the minor bugs that I've been having. Will finish working today at home so I can get on with UI stuff tomorrow in class.