Game Mechanic and Queue finished
Finished working on the game mechanics to a point where minor bugs are no longer a problem and work as I intend it. I also added the queue (the preview that lets you see what pieces are coming up). I just need to add the 'hold' feature and then I can start working on UI.