Commit 6/8
I started working on the UI. Specifically, I worked on making the preview pieces actually line up. Right now, they aren't aligned at each piece's center so it looks ugly. I am going to find the center of the tetrimino so I can shift it so that the preview pieces (which are in a vertical queue) to look more aligned. Almost figured it out -- just need to deal with edge cases where the piece isn't symmetrical.