Commit 6/12
I forgot to commit in class because I was in a hurry to leave. I worked on UI stuff that isn't showing up on the changed files because changing UI stuff around happens inside the scene, which can only be viewed in binary because it's not a format supported by github. Right now, I have implemented the mode selection buttons, and put in the background for the gameplay scene and the title scene. Next time, I plan on adding the pause menu to the gameplay scene.